Decomposed functions in multiple files and fixed warnings
